IGNITE-22807 / make storage an interface
