+++ modules/partition-replicator/build.gradle
@@ -76,6 +76,7 @@ dependencies {
     integrationTestImplementation project(':ignite-placement-driver')
     integrationTestImplementation project(':ignite-affinity')
     integrationTestImplementation project(':ignite-runner')
+    integrationTestImplementation project(':ignite-system-disaster-recovery')
 }
 
 description = 'ignite-partition-replicator'
